+++ lib/fs_rpc.h
@@ -18,19 +18,19 @@
 
 int fs_rpc_mkdir(const char *path, mode_t mode);
 int fs_rpc_inode_create(const char *path, mode_t mode, size_t chunk_size,
-			uint32_t *i_ino);
+			uint64_t *i_ino);
 int fs_rpc_inode_stat(const char *path, fs_stat_t *st);
-int fs_rpc_inode_chunk_stat(uint32_t i_ino, uint32_t index, size_t *size);
-int fs_rpc_inode_truncate(uint32_t i_ino, uint32_t index, off_t offset);
-void *fs_async_rpc_inode_write(uint32_t i_ino, uint32_t index, off_t offset,
+int fs_rpc_inode_chunk_stat(uint64_t i_ino, uint64_t index, size_t *size);
+int fs_rpc_inode_truncate(uint64_t i_ino, uint64_t index, off_t offset);
+void *fs_async_rpc_inode_write(uint64_t i_ino, uint64_t index, off_t offset,
 			       size_t size, const void *buf);
 ssize_t fs_async_rpc_inode_write_wait(void **hdles, int nreqs);
-void *fs_async_rpc_inode_read(uint32_t i_ino, uint32_t index, off_t offset,
+void *fs_async_rpc_inode_read(uint64_t i_ino, uint64_t index, off_t offset,
 			      size_t size, void *buf);
 ssize_t fs_async_rpc_inode_read_wait(void **hdles, int nreqs);
 int fs_rpc_readdir(const char *path, void *arg,
 		   void (*filler)(void *, const char *, const struct stat *));
-int fs_rpc_inode_unlink(const char *path, uint32_t *i_ino);
+int fs_rpc_inode_unlink(const char *path, uint64_t *i_ino);
 int fs_rpc_inode_unlink_all(const char *path);
 int fs_rpc_dir_move(const char *oldpath, const char *newpath);
